+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - MySQL 9: Create routines with LANGUAGE JAVASCRIPT
 - MS SQL: Get the base type of a column declared with a user-defined type
 - ClickHouse: Print only errors from ClickHouse (GHSA-77qq-q8fv-x45v, regression from 6.0.0)
+- ClickHouse: Highlight the queries
+- Firebird: Highlight the queries
 ### Plugins
 - New method assetUrl() returning the URL of a static file in the compiled Adminer